Fix #40 and Fix #32
