+++ component/src/main/java/org/wso2/extension/siddhi/io/http/sink/HttpSink.java
@@ -635,7 +635,6 @@ public class HttpSink extends Sink {
 
                 break;
             }
-
         }
         return httpCarbonMessage;
     }
