+++ component/src/main/java/org/wso2/extension/siddhi/io/http/source/HttpWorkerThread.java
@@ -74,7 +74,7 @@ public class HttpWorkerThread implements Runnable {
             try {
                 buf.close();
             } catch (IOException e) {
-                logger.error("Error closing byte buf");
+                logger.error("Error closing byte buf", e);
             }
         }
     }
