+++ component/src/test/java/org/wso2/extension/siddhi/io/http/source/HttpSourceBasicAuthTestCase.java
@@ -22,8 +22,6 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import org.wso2.extension.siddhi.io.http.source.util.HttpTestUtil;
-import org.wso2.extension.siddhi.map.json.sourcemapper.JsonSourceMapper;
-import org.wso2.extension.siddhi.map.text.sourcemapper.TextSourceMapper;
 import org.wso2.extension.siddhi.map.xml.sourcemapper.XmlSourceMapper;
 import org.wso2.siddhi.core.SiddhiAppRuntime;
 import org.wso2.siddhi.core.SiddhiManager;
@@ -40,11 +38,11 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * Test cases for mapping types.
+ * Test cases for basic authentication.
  */
-public class HttpSourceMappingTestCases {
+public class HttpSourceBasicAuthTestCase {
     private static final org.apache.log4j.Logger logger = org.apache.log4j.Logger.getLogger
-            (HttpSourceMappingTestCases.class);
+            (HttpSourceBasicAuthTestCase.class);
     private AtomicInteger eventCount = new AtomicInteger(0);
     private int waitTime = 50;
     private int timeout = 30000;
@@ -55,20 +53,145 @@ public class HttpSourceMappingTestCases {
     }
 
     /**
-     * Creating test for publishing events with XML mapping.
+     * Creating test for publishing events with basic auth false.
      * @throws Exception Interrupted exception
      */
     @Test
-    public void testXmlMapping() throws Exception {
-        logger.info("Creating test for publishing events with XML mapping.");
+    public void testBasicAuthFalse() throws Exception {
+        logger.info(" Creating test for publishing events with basic auth false.");
         URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
         List<String> receivedEventNameList = new ArrayList<>(2);
         PersistenceStore persistenceStore = new InMemoryPersistenceStore();
         SiddhiManager siddhiManager = new SiddhiManager();
         siddhiManager.setPersistenceStore(persistenceStore);
         siddhiManager.setExtension("xml-input-mapper", XmlSourceMapper.class);
-        String inStreamDefinition = "" + "@source(type='http', @map(type='xml'), "
-                + "receiver.url='http://localhost:8005/endpoints/RecPro', " + "basic.auth.enabled='false'" + ")"
+        String inStreamDefinition = "@source(type='http', @map(type='xml'), receiver.url='http://localhost:8005"
+                + "/endpoints/RecPro', basic.auth.enabled='false' )"
+                + "define stream inputStream (name string, age int, country string);";
+        String query = (
+                "@info(name = 'query') "
+                        + "from inputStream "
+                        + "select * "
+                        + "insert into outputStream;"
+                        );
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager
+                .createSiddhiAppRuntime(inStreamDefinition + query);
+
+        siddhiAppRuntime.addCallback("query", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                for (Event event : inEvents) {
+                    eventCount.incrementAndGet();
+                    receivedEventNameList.add(event.getData(0).toString());
+                }
+            }
+        });
+        siddhiAppRuntime.start();
+        // publishing events
+        List<String> expected = new ArrayList<>(2);
+        expected.add("John");
+        expected.add("Mike");
+        String event1 = "<events>"
+                            + "<event>"
+                               + "<name>John</name>"
+                               + "<age>100</age>"
+                               + "<country>AUS</country>"
+                            + "</event>"
+                        + "</events>";
+        String event2 = "<events>"
+                             + "<event>"
+                               + "<name>Mike</name>"
+                               + "<age>20</age>"
+                               + "<country>USA</country>"
+                             + "</event>"
+                        + "</events>";
+        HttpTestUtil.httpPublishEvent(event1, baseURI, "/endpoints/RecPro",
+                "POST");
+        HttpTestUtil.httpPublishEvent(event2, baseURI, "/endpoints/RecPro",
+                "POST");
+        SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
+        Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
+        siddhiAppRuntime.shutdown();
+    }
+
+    /**
+     * Creating test for publishing events with basic auth false.
+     * @throws Exception Interrupted exception
+     */
+    @Test
+    public void testBasicAuthTrue() throws Exception {
+        logger.info(" Creating test for publishing events with basic auth false.");
+        URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
+        List<String> receivedEventNameList = new ArrayList<>(2);
+        PersistenceStore persistenceStore = new InMemoryPersistenceStore();
+        SiddhiManager siddhiManager = new SiddhiManager();
+        siddhiManager.setPersistenceStore(persistenceStore);
+        siddhiManager.setExtension("xml-input-mapper", XmlSourceMapper.class);
+        String inStreamDefinition = "@source(type='http', @map(type='xml'), receiver.url='http://localhost:8005"
+                + "/endpoints/RecPro', basic.auth.enabled='true' )"
+                + "define stream inputStream (name string, age int, country string);";
+        String query = (
+                "@info(name = 'query') "
+                        + "from inputStream "
+                        + "select * "
+                        + "insert into outputStream;"
+        );
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager
+                .createSiddhiAppRuntime(inStreamDefinition + query);
+
+        siddhiAppRuntime.addCallback("query", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                for (Event event : inEvents) {
+                    eventCount.incrementAndGet();
+                    receivedEventNameList.add(event.getData(0).toString());
+                }
+            }
+        });
+        siddhiAppRuntime.start();
+        // publishing events
+        List<String> expected = new ArrayList<>(2);
+        expected.add("John");
+        expected.add("Mike");
+        String event1 = "<events>"
+                + "<event>"
+                + "<name>John</name>"
+                + "<age>100</age>"
+                + "<country>AUS</country>"
+                + "</event>"
+                + "</events>";
+        String event2 = "<events>"
+                + "<event>"
+                + "<name>Mike</name>"
+                + "<age>20</age>"
+                + "<country>USA</country>"
+                + "</event>"
+                + "</events>";
+        HttpTestUtil.httpPublishEvent(event1, baseURI, "/endpoints/RecPro",
+                "POST");
+        HttpTestUtil.httpPublishEvent(event2, baseURI, "/endpoints/RecPro",
+                "POST");
+        SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
+        Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
+        siddhiAppRuntime.shutdown();
+    }
+    /**
+     * Creating test for publishing events with basic auth is not provided.
+     * @throws Exception Interrupted exception
+     */
+    @Test
+    public void testBasicAuthNotProvided() throws Exception {
+        logger.info("Creating test for publishing events with basic auth is not provided.");
+        URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
+        List<String> receivedEventNameList = new ArrayList<>(2);
+        PersistenceStore persistenceStore = new InMemoryPersistenceStore();
+        SiddhiManager siddhiManager = new SiddhiManager();
+        siddhiManager.setPersistenceStore(persistenceStore);
+        siddhiManager.setExtension("xml-input-mapper", XmlSourceMapper.class);
+        String inStreamDefinition = "@source(type='http', @map(type='xml'), receiver.url='http://localhost:8005"
+                + "/endpoints/RecPro' )"
                 + "define stream inputStream (name string, age int, country string);";
         String query = (
                 "@info(name = 'query') "
@@ -116,179 +239,5 @@ public class HttpSourceMappingTestCases {
         Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
         siddhiAppRuntime.shutdown();
     }
-    /**
-     * Creating test for publishing events with Text mapping.
-     *
-     * @throws Exception Interrupted exception
-     */
-    @Test
-    public void testTextMappingSingle() throws Exception {
-        AtomicInteger eventCount = new AtomicInteger(0);
-        int waitTime = 50;
-        int timeout = 30000;
-        URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
-        List<String> receivedEventNameList = new ArrayList<>(2);
-        PersistenceStore persistenceStore = new InMemoryPersistenceStore();
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setPersistenceStore(persistenceStore);
-        siddhiManager.setExtension("text", TextSourceMapper.class);
-        String inStreamDefinition = "" + "@source(type='http',  @map(type='text'), "
-                + "receiver.url='http://localhost:8005/endpoints/RecPro', " + "basic.auth.enabled='false'" + ")"
-                + "define stream inputStream (name string, age int, country string);";
-        String query = (
-                "@info(name = 'query') "
-                        + "from inputStream "
-                        + "select *  "
-                        + "insert into outputStream;"
-        );
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager
-                .createSiddhiAppRuntime(inStreamDefinition + query);
-
-        siddhiAppRuntime.addCallback("query", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timeStamp, inEvents, removeEvents);
-                for (Event event : inEvents) {
-                    eventCount.incrementAndGet();
-                    receivedEventNameList.add(event.getData(0).toString());
-                }
-            }
-        });
-        siddhiAppRuntime.start();
-
-        // publishing events
-        List<String> expected = new ArrayList<>(2);
-        expected.add("John");
-        expected.add("Mike");
-        String event1 = "name:\'John\',\n" +
-                "age:100,\n" +
-                "country:\'USA\'";
-        String event2 = "name:\'Mike\',\n" +
-                "age:100,\n" +
-                "country:\'USA\'";
-        HttpTestUtil.httpPublishEvent(event1, baseURI, "/endpoints/RecPro",
-                "POST");
-        HttpTestUtil.httpPublishEvent(event2, baseURI, "/endpoints/RecPro",
-                "POST");
-        SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
-        Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
-        siddhiAppRuntime.shutdown();
-    }
-    /**
-     * Creating test for publishing events with Text mapping.
-     * @throws Exception Interrupted exception
-     */
-    @Test
-    public void testTextMapping() throws Exception {
-        logger.info("Creating test for publishing events with Text mapping.");
-        URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
-        List<String> receivedEventNameList = new ArrayList<>(2);
-        PersistenceStore persistenceStore = new InMemoryPersistenceStore();
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setPersistenceStore(persistenceStore);
-        siddhiManager.setExtension("text", TextSourceMapper.class);
-        String inStreamDefinition = "" + "@source(type='http',  @map(type='text',fail.on.missing.attribute='false'), "
-                + "receiver.url='http://localhost:8005/endpoints/RecPro', " + "basic.auth.enabled='false'" + ")"
-                + "define stream inputStream (name string, age int, country string);";
-        String query = (
-                "@info(name = 'query') "
-                        + "from inputStream "
-                        + "select *  "
-                        + "insert into outputStream;"
-                    );
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager
-                .createSiddhiAppRuntime(inStreamDefinition + query);
-
-        siddhiAppRuntime.addCallback("query", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timeStamp, inEvents, removeEvents);
-                for (Event event : inEvents) {
-                    eventCount.incrementAndGet();
-                    receivedEventNameList.add(event.getData(0).toString());
-                }
-            }
-        });
-        siddhiAppRuntime.start();
-
-        // publishing events
-        List<String> expected = new ArrayList<>(2);
-        expected.add("John");
-        expected.add("Mike");
-        String event1 = "name:\"John\",\n" +
-                "age:100,\n" +
-                "country:\"USA\"";
-        String event2 = "name:\"Mike\",\n" +
-                "age:100,\n" +
-                "country:\"USA\"";
-        HttpTestUtil.httpPublishEvent(event1, baseURI, "/endpoints/RecPro",
-                "POST");
-        HttpTestUtil.httpPublishEvent(event2, baseURI, "/endpoints/RecPro",
-                "POST");
-        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
-        Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
-        siddhiAppRuntime.shutdown();
-    }
-
-    /**
-     * Creating test for publishing events with Json mapping.
-     * @throws Exception Interrupted exception
-     */
-    @Test
-    public void testJsonMapping() throws Exception {
-        logger.info("Creating test for publishing events with Json mapping.");
-        URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8005));
-        List<String> receivedEventNameList = new ArrayList<>(2);
-        PersistenceStore persistenceStore = new InMemoryPersistenceStore();
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setPersistenceStore(persistenceStore);
-        siddhiManager.setExtension("xml-input-mapper", JsonSourceMapper.class);
-        String inStreamDefinition = "" + "@source(type='http', @map(type='json'), "
-                + "receiver.url='http://localhost:8005/endpoints/RecPro', " + "basic.auth.enabled='false'" + ")"
-                + "define stream inputStream (name string, age int, country string);";
-        String query = ("@info(name = 'query') "
-                + "from inputStream "
-                + "select *  "
-                + "insert into outputStream;"
-                );
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager
-                .createSiddhiAppRuntime(inStreamDefinition + query);
-
-        siddhiAppRuntime.addCallback("query", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timeStamp, inEvents, removeEvents);
-                for (Event event : inEvents) {
-                    eventCount.incrementAndGet();
-                    receivedEventNameList.add(event.getData(0).toString());
-                }
-            }
-        });
-        siddhiAppRuntime.start();
-        // publishing events
-        List<String> expected = new ArrayList<>(2);
-        expected.add("John");
-        expected.add("Mike");
-        String event1 = " {\n" +
-                "      \"event\":{\n" +
-                "         \"name\":\"John\",\n" +
-                "         \"age\":55,\n" +
-                "         \"country\":\"US\"\n" +
-                "      }\n" +
-                " }";
-        String event2 = " {\n" +
-                "      \"event\":{\n" +
-                "         \"name\":\"Mike\",\n" +
-                "         \"age\":55,\n" +
-                "         \"country\":\"US\"\n" +
-                "      }\n" +
-                " }";
-        HttpTestUtil.httpPublishEvent(event1, baseURI, "/endpoints/RecPro",
-                "POST");
-        HttpTestUtil.httpPublishEvent(event2, baseURI, "/endpoints/RecPro",
-                "POST");
-        SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
-        Assert.assertEquals(receivedEventNameList.toString(), expected.toString());
-        siddhiAppRuntime.shutdown();
-    }
 }
+
