+++ component/src/test/java/org/wso2/extension/siddhi/io/http/source/util/HttpServerUtil.java
@@ -62,7 +62,7 @@ class HttpServerUtil {
     }
 
     static void handleException(IOException ex) {
-        log.error("IOException occurred while running the HttpsSourceTestCaseForSSL", ex);
+        log.error("IOException occurred while running the HttpsSSLSourceTestCase", ex);
     }
 }
